Dinning Page
created dinning.html, linked the index page with the dinning page, created a dinner.css file, most of the contents are ready...i will clean it up later. Don't valiate this html one ID used three times of once is some of the errors. but will fix all that later. on another branch